macros: new DECL_CONST macro
This macro is needed where a constant value can be used in both assembly
and C code, but it needs a specific type in C code for correct results.
This type is usually assigned with a suffix such as "UL", which makes
assembly fail. DECL_CONST creates different forms based on the language.