x86/pmap: fix pmap_kgrow
The limit of the kernel pmap is aligned to the size mapped by a page directory
entry (second level of the translation hierarchy), but the code assumes it's
aligned to the size mapped by a root entry (the highest level of the
translation hierarchy). Fix this by aligning addresses instead of incrementing
them with an offset.