kern/thread: never fail when adding a thread to a run queue
Handling this kind of error makes threading hazardous. It's just way better
to simply consider a thread can always be added to a run queue. The reason
it could fail is that the time-sharing scheduling algorithm doesn't support
a global weight greater than 2^32. A direct implication is that, because of
this limitation, the maximum number of threads should be limited too.