x86/param: update kernel space end address on i386
The kernel image is loaded where the kernel space ends. The previous end
address left 2 MiB to the kernel image. Since boot stacks are now statically
allocated, a kernel configured for a large number of processors would
overflow the area reserved for the kernel image. Reduce the kernel space
end address so that 64 MiB are now available for the kernel image.