x86/pmap: replace pmap_kprotect with pmap_protect
This change is merely a slight interface modification to get rid of a
function that shouldn't be exported and replace it with the true entry
point for setting physical mapping protection.

The new function isn't complete yet and cannot handle physical maps that
aren't the kernel pmap or the currently loaded pmap.