vm/vm_kmem: fix page allocation error handling
Although not yet clearly stated since it's not yet decided whether it will
remain the normal behaviour, the pmap functions assume physical mappings
to be valid on removal or protection update. But the kernel virtual memory
allocation function vm_kmem_alloc would merely call vm_kmem_free if a page
couldn't be allocated, resulting on the unconditional removal of the
underlying physical mappings of the whole range, including addresses for
which no mapping was created.