x86/pmap: fix initialization of percpu variables
Application processors should never initialize data that are shared with
other processors. In this case, it leaded to a race which could make
pmap_update access uninitialized percpu data. But, since percpu areas
are initially copies of the percpu section (which is the percpu area of
the main processor), uninitialized percpu variables would actually have
valid values, those inherited from the main processor. This would result
in hard-to-debug data corruption.