x86/pmap: directmap update
This module is probably the most impacted by the direct physical
mapping.

First, it establishes the direct physical mapping before paging is
enabled, and uses the largest available page size when doing so.

In addition, the recursive mapping of PTEs is removed, since all
page table pages are allocated from the direct physical mapping.
This changes the way mapping requests are processed.

The ability to access any page table page at any time removes the
need for virtual addresses reserved for temporary mappings.

Since it's now perfectly possible to return physical address 0 for
a kernel mapping, change the interface of the pmap_extract function
and rename it to pmap_kextract to stress the fact that it should
only be used for kernel mappings.