x86/pmap: rework PAE
Instead of presenting PAE mode as a two-level hierarchy (the same as
non-PAE 32-bits) with a page directory 4 pages wide, present it as a
three-level hierarchy. The purpose of this change is to break the
requirement of the page directory being 4 contiguous pages, which
is likely to fail because of fragmentation.