kern/init: fix LTO builds
Initialization operations are never directly referenced, as they're
retrieved by accessing the .init.ops subsection. As a result, whole
program optimizations as applied during link time optimizations cause
the structures and most of the related code not to be compiled in.

Use the recently added __used attribute macro to fix this issue.