tools/build_configs: rework and improve
The most important change is the addition of command line arguments used
to select options sets. The purpose of options sets selection is to make
continuous integration quick for the most common options, but still have
a way to test much larger sets of options, e.g. with nightly builds.