kern/sleepq: fix signal behaviour
In order to correctly implement the behaviour of condition_signal, which
must wake up at least one waiting thread if there's any, sleepq_signal
wakes up the oldest waiter and shifts it.

The FIFO behaviour is removed from the specification, as the only user
which relied on it was a previous condition variable implementation.