kern/atomic: fix consume memory order
The consume memory order is known to be problematic in the C11 specification,
and as a result, most compilers alias it to acquire. Based on the assumption
that the kernel doesn't run on architectures with an Alpha-like memory
model, the consume memory order was aliased to relaxed, letting the CPU
identify data dependencies. But there may also be issues at compile time,
making this hack is dangerous. As a result, the actual consume memory order
as implemented by the compiler is now used to define the kernel consume
memory order.

See https://lwn.net/Articles/588300/.