kern/panic: remove initialization operation
Maintaining correct dependencies for the panic function is too
complicated. It may be used very early, but would depend on the
console, which is normally available quite late during initialization.

Instead, the behavior of panic is now "best effort", i.e. if called at
a time where it works, good, otherwise, the behavior is undefined,
but should result in a freeze or a reset most of the time.